Skip set.seed() if NA_integer_ is given.
This should address #52 but currently it does not work because
`NA_integer_` is implemented as the maximum negative value, and
a negative seed causes an error "negative length vectors are not allowed".
`base::set.seed()` accepts negative seeds without any problem,
which indicates ClusterR is doing something wrong with negative seeds.